Revert "SystemUI: more keyguard visualizer tweaks"
This reverts commit b821a45ae5481e22c4909551310080768bc13d00.

+++ packages/SystemUI/src/com/android/systemui/statusbar/BackDropView.java
@@ -25,7 +25,6 @@ import android.graphics.Color;
 import android.graphics.DashPathEffect;
 import android.graphics.Paint;
 import android.graphics.Rect;
-import android.os.AsyncTask;
 import android.support.v7.graphics.Palette;
 import android.util.AttributeSet;
 import android.view.View;
@@ -40,7 +39,7 @@ import com.pheelicks.visualizer.renderer.Renderer;
 /**
  * A view who contains media artwork.
  */
-public class BackDropView extends FrameLayout implements Palette.PaletteAsyncListener {
+public class BackDropView extends FrameLayout {
     private Runnable mOnVisibilityChangedRunnable;
 
     public BackDropView(Context context) {
@@ -82,7 +81,6 @@ public class BackDropView extends FrameLayout implements Palette.PaletteAsyncLis
     private boolean mAnimating = false;
     private boolean mTouching = false;
     private boolean mPowerSaveMode = false;
-    private boolean mQsExpanded = false;
     private int mColor;
 
     private VisualizerView mVisualizer;
@@ -92,8 +90,10 @@ public class BackDropView extends FrameLayout implements Palette.PaletteAsyncLis
     private final Runnable mLinkVisualizer = new Runnable() {
         @Override
         public void run() {
-            if (mVisualizer != null) {
+            if (mVisualizer != null && !mLinked) {
                 mVisualizer.link(0);
+                mVisualizer.animate().alpha(1f).setDuration(300);
+                mLinked = true;
             }
         }
     };
@@ -101,8 +101,10 @@ public class BackDropView extends FrameLayout implements Palette.PaletteAsyncLis
     private final Runnable mUnlinkVisualizer = new Runnable() {
         @Override
         public void run() {
-            if (mVisualizer != null) {
+            if (mVisualizer != null && mLinked) {
+                mVisualizer.animate().alpha(0f).setDuration(0);
                 mVisualizer.unlink();
+                mLinked = false;
             }
         }
     };
@@ -165,7 +167,7 @@ public class BackDropView extends FrameLayout implements Palette.PaletteAsyncLis
         paint.setStrokeWidth(res.getDimensionPixelSize(R.dimen.kg_visualizer_path_stroke_width));
         paint.setAntiAlias(true);
         paint.setColor(mColor);
-        paint.setPathEffect(new DashPathEffect(new float[]{
+        paint.setPathEffect(new DashPathEffect(new float[] {
                 res.getDimensionPixelSize(R.dimen.kg_visualizer_path_effect_1),
                 res.getDimensionPixelSize(R.dimen.kg_visualizer_path_effect_2)
         }, 0));
@@ -216,43 +218,24 @@ public class BackDropView extends FrameLayout implements Palette.PaletteAsyncLis
         }
     }
 
-    public void setQsExpanded(boolean qsExpanded) {
-        if (mQsExpanded != qsExpanded) {
-            mQsExpanded = qsExpanded;
-            checkStateChanged();
-        }
-    }
-
-    @Override
-    public void onGenerated(Palette palette) {
+    public void setBitmap(Bitmap bitmap) {
         int color = Color.TRANSPARENT;
 
-        color = palette.getLightVibrantColor(color);
+        Palette p = Palette.generate(bitmap);
+        color = p.getVibrantColor(color);
         if (color == Color.TRANSPARENT) {
-            color = palette.getVibrantColor(color);
+            color = p.getLightVibrantColor(color);
             if (color == Color.TRANSPARENT) {
-                color = palette.getDarkVibrantColor(color);
+                color = p.getDarkVibrantColor(color);
+                if (color == Color.TRANSPARENT) {
+                    color = mContext.getResources().getColor(R.color.equalizer_fill_color);
+                }
             }
         }
-        setVisualizerColor(color, true, true);
-    }
 
-    public void setBitmap(Bitmap bitmap) {
-        if (bitmap != null) {
-            Palette.generateAsync(bitmap, this);
-        } else {
-            setVisualizerColor(Color.TRANSPARENT, false, false);
-        }
-    }
-
-    private void setVisualizerColor(int color, boolean forceColor, boolean animate) {
-        if (color == Color.TRANSPARENT && forceColor) {
-            color = mContext.getResources().getColor(R.color.equalizer_fill_color);
-        }
         if (mColor != color) {
             mColor = color;
-
-            if (mLinked && animate) {
+            if (mLinked) {
                 if (mVisualizerColorAnimator != null) {
                     mVisualizerColorAnimator.cancel();
                 }
@@ -268,18 +251,10 @@ public class BackDropView extends FrameLayout implements Palette.PaletteAsyncLis
     }
 
     private void checkStateChanged() {
-        if (mVisible && mPlaying && !mAnimating && !mTouching && !mPowerSaveMode && !mQsExpanded) {
-            if (!mLinked) {
-                mVisualizer.animate().alpha(1f).setDuration(300);
-                AsyncTask.execute(mLinkVisualizer);
-                mLinked = true;
-            }
+        if (mVisible && mPlaying && !mAnimating && !mTouching && !mPowerSaveMode) {
+            mLinkVisualizer.run();
         } else {
-            if (mLinked) {
-                mVisualizer.animate().alpha(0f).setDuration(0);
-                AsyncTask.execute(mUnlinkVisualizer);
-                mLinked = false;
-            }
+            mUnlinkVisualizer.run();
         }
     }
 }
